requirements.txt
bcrypt==4.2.1
cffi==1.17.1
cryptography==43.0.3
future==1.0.0
netmiko==4.4.0
ntc_templates==7.4.0
paramiko==3.5.0
pycparser==2.22
PyNaCl==1.5.0
pyserial==3.5
python-dotenv==1.0.1
PyYAML==6.0.2
scp==0.15.0
setuptools==75.6.0
six==1.16.0
textfsm==1.1.3
